+++ src/requires/hudclass.lua
@@ -20,6 +20,16 @@ function HudClass(Nav, c, u, s, atlas, radar_1, radar_2, antigrav, hover, shield
     local maxBrakeTime = 0
     local damageMessage = ""
     local WeaponPanelID = nil
+    local PrimaryR = SafeR
+    local PrimaryG = SafeG
+    local PrimaryB = SafeB
+    local rgb = [[rgb(]] .. mfloor(PrimaryR + 0.5) .. "," .. mfloor(PrimaryG + 0.5) .. "," .. mfloor(PrimaryB + 0.5) .. [[)]]
+    local rgbdim = [[rgb(]] .. mfloor(PrimaryR * 0.9 + 0.5) .. "," .. mfloor(PrimaryG * 0.9 + 0.5) .. "," ..   mfloor(PrimaryB * 0.9 + 0.5) .. [[)]]
+    local totalDistanceTrip = 0
+    local flightTime = 0
+    local lastOdometerOutput = ""
+    local lastTravelTime = systime()
+    local repairArrows = false
 
     --Local Huds Functions
         -- safezone() variables
@@ -852,7 +862,9 @@ function HudClass(Nav, c, u, s, atlas, radar_1, radar_2, antigrav, hover, shield
             end
             if BrakeLanding then
                 if StrongBrakes then
-                    newContent[#newContent + 1] = svgText(warningX, apY, "Brake-Landing", "warnings")
+                    local str = "Brake-Landing"
+                    if alignHeading then str = str..": Aligning" end
+                    newContent[#newContent + 1] = svgText(warningX, apY, str, "warnings")
                 else
                     newContent[#newContent + 1] = svgText(warningX, apY, "Coast-Landing", "warnings")
                 end
@@ -1065,7 +1077,7 @@ function HudClass(Nav, c, u, s, atlas, radar_1, radar_2, antigrav, hover, shield
                 addTable(help, helpSpace)
                 if shield_1 then
                     table.insert(help,"Alt-Shift-6: Vent shields")
-                    table.insert(help,"Alt-Shift-7: Toggle shied off/on")
+                    if not AutoShieldToggle then table.insert(help,"Alt-Shift-7: Toggle shield off/on") end
                 end
             end
             if CustomTarget ~= nil then
@@ -2200,16 +2212,10 @@ function HudClass(Nav, c, u, s, atlas, radar_1, radar_2, antigrav, hover, shield
             PrimaryR = PvPR
             PrimaryG = PvPG
             PrimaryB = PvPB
-            if shield_1 and AutoShieldToggle and shield_1.getState() == 0 then
-                shield_1.toggle()
-            end
         else
             PrimaryR = SafeR
             PrimaryG = SafeG
             PrimaryB = SafeB
-            if shield_1 and AutoShieldToggle and shield_1.getState() == 1 then
-                shield_1.toggle()
-            end
         end
         rgb = [[rgb(]] .. mfloor(PrimaryR + 0.6) .. "," .. mfloor(PrimaryG + 0.6) .. "," .. mfloor(PrimaryB + 0.6) .. [[)]]
         rgbdim = [[rgb(]] .. mfloor(PrimaryR * 0.8 + 0.5) .. "," .. mfloor(PrimaryG * 0.8 + 0.5) .. "," ..   mfloor(PrimaryB * 0.8 + 0.5) .. [[)]]    
@@ -2583,7 +2589,7 @@ function HudClass(Nav, c, u, s, atlas, radar_1, radar_2, antigrav, hover, shield
         local msg, where
 
     function Hud.DrawRadarInfo()
-        radarMessage = RADAR.GetRadarHud(friendx, friendy, radarX, radarY)
+        radarMessage = RADAR.GetRadarHud(friendx, friendy, radarX, radarY) 
     end
 
     function Hud.DrawTanks()
@@ -2604,7 +2610,6 @@ function HudClass(Nav, c, u, s, atlas, radar_1, radar_2, antigrav, hover, shield
         local resistances = shield_1.getResistances()
         local resistString = "A: "..(10+resistances[1]*100).."% / E: "..(10+resistances[2]*100).."% / K:"..(10+resistances[3]*100).."% / T: "..(10+resistances[4]*100).."%"
         local x, y = shieldX -60, shieldY+30
-        local shieldPercent = mfloor(0.5 + shield_1.getShieldHitpoints() * 100 / shield_1.getMaxShieldHitpoints())
         local colorMod = mfloor(shieldPercent * 2.55)
         local color = stringf("rgb(%d,%d,%d)", 255 - colorMod, colorMod, 0)
         local class = ""